requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
aiosqlite
python-jose
bcrypt
minio
python-multipart
greenlet
